Deaths and Garbage
I felt it was a bit too weird of a design to have the player get an ending with every single path so now you can die! This will probably just add to your death counte, and the player will have to reset. Imma gonna work on the one garabage path I forgot about where you land in a landfill. But m going to work on the read me